conform: fix checking of a custom data type
A custom data type represented as a module name in the
conform schema. Previously we checked a type with the call of the
Code.ensure_loaded/1, but it returns {:error, :embedded} if an
application release launched in 'foreground' mode. The problem that
we can't load any modules in runtime if the erlang VM was launched
in embeded mode.

Now we are using Code.ensure_load?/1 which does not try to load a
given module, but just checks was it loaded or not.